+++ operator_ui/__tests__/actions.test.ts
@@ -3,7 +3,7 @@ import * as models from 'core/store/models'
 import { bindActionCreators, Middleware } from 'redux'
 import configureStore from 'redux-mock-store'
 import thunk from 'redux-thunk'
-import * as actionCreators from '../src/actions'
+import * as actionCreators from '../src/actionCreators'
 import { ResourceActionType } from '../src/reducers/actions'
 import jsonApiJobSpecFactory from '../support/factories/jsonApiJobSpec'
 import jsonApiJobSpecRunFactory from '../support/factories/jsonApiJobSpecRun'
