+++ operator_ui/src/containers/Bridges/Edit.js
@@ -10,7 +10,7 @@ import Divider from '@material-ui/core/Divider'
 import ErrorMessage from 'components/Notifications/DefaultError'
 import Button from 'components/Button'
 import bridgeSelector from 'selectors/bridge'
-import { fetchBridgeSpec, updateBridge } from 'actions'
+import { fetchBridgeSpec, updateBridge } from 'actionCreators'
 import matchRouteAndMapDispatchToProps from 'utils/matchRouteAndMapDispatchToProps'
 import BaseLink from 'components/BaseLink'
 import Content from 'components/Content'
