+++ operator_ui/src/containers/Bridges/New.tsx
@@ -8,7 +8,7 @@ import Typography from '@material-ui/core/Typography'
 import Form from 'components/Bridges/Form'
 import ErrorMessage from 'components/Notifications/DefaultError'
 import Content from 'components/Content'
-import { createBridge } from 'actions'
+import { createBridge } from 'actionCreators'
 import matchRouteAndMapDispatchToProps from 'utils/matchRouteAndMapDispatchToProps'
 import BaseLink from '../../components/BaseLink'
 import { BridgeTypeAuthentication } from 'core/store/models'
