+++ operator_ui/src/containers/Bridges/Show.tsx
@@ -3,7 +3,7 @@ import CardContent from '@material-ui/core/CardContent'
 import Divider from '@material-ui/core/Divider'
 import Grid from '@material-ui/core/Grid'
 import Typography from '@material-ui/core/Typography'
-import { fetchBridgeSpec } from 'actions'
+import { fetchBridgeSpec } from 'actionCreators'
 import BaseLink from 'components/BaseLink'
 import Button from 'components/Button'
 import Content from 'components/Content'
