+++ operator_ui/src/containers/Configuration/DeleteJobRuns.js
@@ -4,7 +4,7 @@ import Grid from '@material-ui/core/Grid'
 import Slide from '@material-ui/core/Slide'
 import { withStyles } from '@material-ui/core/styles'
 import Typography from '@material-ui/core/Typography'
-import { deleteCompletedJobRuns, deleteErroredJobRuns } from 'actions'
+import { deleteCompletedJobRuns, deleteErroredJobRuns } from 'actionCreators'
 import Button from 'components/Button'
 import React, { useState } from 'react'
 import { connect } from 'react-redux'
