+++ operator_ui/src/containers/Header.tsx
@@ -21,7 +21,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 import ReactResizeDetector from 'react-resize-detector'
 import { bindActionCreators, Dispatch } from 'redux'
-import { submitSignOut } from '../actions'
+import { submitSignOut } from 'actionCreators'
 import AvatarMenu from '../components/AvatarMenu'
 import BaseLink from '../components/BaseLink'
 import LoadingBar from '../components/LoadingBar'
