+++ operator_ui/src/containers/JobRuns/Show/Overview.js
@@ -4,7 +4,7 @@ import { withStyles } from '@material-ui/core/styles'
 import Grid from '@material-ui/core/Grid'
 import Card from '@material-ui/core/Card'
 import matchRouteAndMapDispatchToProps from 'utils/matchRouteAndMapDispatchToProps'
-import { fetchJobRun } from 'actions'
+import { fetchJobRun } from 'actionCreators'
 import jobRunSelector from 'selectors/jobRun'
 import Content from 'components/Content'
 import StatusCard from 'components/JobRuns/StatusCard'
