+++ operator_ui/src/containers/Jobs/Definition.js
@@ -9,7 +9,7 @@ import Divider from '@material-ui/core/Divider'
 import Content from 'components/Content'
 import PrettyJson from 'components/PrettyJson'
 import RegionalNav from './RegionalNav'
-import { fetchJob, createJobRun } from 'actions'
+import { fetchJob, createJobRun } from 'actionCreators'
 import jobSelector from 'selectors/job'
 import matchRouteAndMapDispatchToProps from 'utils/matchRouteAndMapDispatchToProps'
 import jobSpecDefinition from 'utils/jobSpecDefinition'
