+++ operator_ui/src/containers/Jobs/RegionalNav.tsx
@@ -16,7 +16,7 @@ import { JobSpec } from 'operator_ui'
 import React from 'react'
 import { connect } from 'react-redux'
 import { Redirect, useRouteMatch } from 'react-router-dom'
-import { createJobRun, deleteJobSpec, fetchJobRuns } from '../../actions'
+import { createJobRun, deleteJobSpec, fetchJobRuns } from 'actionCreators'
 import BaseLink from '../../components/BaseLink'
 import Button from '../../components/Button'
 import CopyJobSpec from '../../components/CopyJobSpec'
