+++ operator_ui/src/containers/SignIn.js
@@ -9,7 +9,7 @@ import Typography from '@material-ui/core/Typography'
 import TextField from '@material-ui/core/TextField'
 import { Grid } from '@material-ui/core'
 import { hot } from 'react-hot-loader'
-import { submitSignIn } from 'actions'
+import { submitSignIn } from 'actionCreators'
 import { renderNotification } from 'containers/Notifications'
 import HexagonLogo from 'components/Logos/Hexagon'
 import matchRouteAndMapDispatchToProps from 'utils/matchRouteAndMapDispatchToProps'
