+++ operator_ui/src/containers/SignOut.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react'
 import PropTypes from 'prop-types'
-import { receiveSignoutSuccess } from 'actions'
+import { receiveSignoutSuccess } from 'actionCreators'
 import { connect } from 'react-redux'
 
 export const SignOut = ({ receiveSignoutSuccess }) => {
